+++ frontend/src/components/games/components/championImg.js
