+++ frontend/src/components/games/components/gameTitle.js
