+++ frontend/src/components/games/components/victory.js
