+++ frontend/src/components/games/game.js
@@ -1,14 +1,25 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
+import Titles from "./components/gameTitle";
+import ChampionDetails from "./components/championDetails";
 import Select from "react-select";
-import Victory from "./victory";
-import ChampionImg from "./championImg";
-import Config from "../configs/config";
-import { saveGamesPlayed, saveTries, saveFirstTries } from "./saveStats";
-import { Reroll } from "./reroll";
+import Victory from "./components/victory";
+import {
+  saveGamesPlayed,
+  saveTries,
+  saveFirstTries,
+} from "../../utils/saveStats";
+import Config from "../../configs/config";
+import { Reroll } from "../../utils/reroll";
+import { Helmet } from "react-helmet";
 import LazyLoad from "react-lazy-load";
-import { SelectStyles, customFilterOption, SelectTheme } from "./selectStyles";
 import { useSelector } from "react-redux";
+
+import {
+  SelectStyles,
+  customFilterOption,
+  SelectTheme,
+} from "./styles/selectStyles";
 
 export default function Game() {
   const [validGuesses, setValidGuesses] = useState([]);
@@ -16,24 +27,18 @@ export default function Game() {
   const [guesses, setGuesses] = useState([]);
   const [currentGuess, setGuess] = useState(validGuesses[0]);
   const [correctGuess, setCorrectGuess] = useState(false);
-  const [sprite, setSprite] = useState("");
   const [title, setTitle] = useState("");
 
   const isColorBlindMode = useSelector(
     (state) => state.colorBlindReducer.isColorBlindMode
   );
 
-  const isMonochrome = useSelector(
-    (state) => state.monochromeReducer.isMonochrome
-  );
-
-  const randomRotate = useSelector(
-    (state) => state.randomRotateReducer.randomRotate
+  const hideResource = useSelector(
+    (state) => state.hideResourceReducer.hideResource
   );
 
   useEffect(() => {
     FetchChampions();
-    FetchSplashArt();
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   const FetchChampions = () => {
@@ -49,23 +54,6 @@ export default function Game() {
             image: champion.image,
           }));
           setValidGuesses(transformedData);
-        }
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  const FetchSplashArt = () => {
-    axios
-      .get(Config.url + "/splash", {
-        headers: { authorization: "Bearer " + localStorage.getItem("token") },
-      })
-      .then((response) => {
-        if (response.data.status === "success") {
-          if (response.data.result) {
-            setSprite(response.data.result);
-          }
         }
       })
       .catch((error) => {
@@ -89,7 +77,7 @@ export default function Game() {
 
     axios
       .post(
-        Config.url + "/splash",
+        Config.url + "/guess",
         { guess: currentGuess },
         {
           headers: { authorization: "Bearer " + localStorage.getItem("token") },
@@ -101,30 +89,18 @@ export default function Game() {
         }
         saveTries(1);
 
-        const isCorrect = response.data.correctGuess;
-        const key = response.data.championKey;
+        const correct = response.data.correctGuess;
+        const data = response.data.properties;
 
-        setChampions((champions) => [[key, isCorrect], ...champions]);
+        setChampions((champions) => [data, ...champions]);
 
-        const spriteImg = document.getElementById("spriteImg");
-
-        if (isCorrect) {
+        if (correct) {
           if (guesses.length === 0) {
             saveFirstTries();
           }
           saveGamesPlayed();
           setCorrectGuess(true);
           setTitle(response.data.title);
-
-          spriteImg.style.filter = "";
-        } else {
-          let blurVal = parseFloat(spriteImg.style.filter.substring(5, 8));
-
-          blurVal -= blurVal * 0.4;
-
-          spriteImg.style.filter = `blur(${blurVal.toString()}em) ${
-            isMonochrome ? "grayscale(1)" : ""
-          }`;
         }
       })
       .catch((error) => {
@@ -134,11 +110,6 @@ export default function Game() {
   };
 
   const Restart = () => {
-    const spriteImg = document.getElementById("spriteImg");
-    spriteImg.style.filter = `blur(1.0em) ${
-      isMonochrome ? "grayscale(1)" : ""
-    }`;
-    FetchSplashArt();
     FetchChampions();
 
     setGuesses([]);
@@ -149,28 +120,21 @@ export default function Game() {
 
   return (
     <div className="container main pt-4 pb-5 mb-5">
-      <h3 className="text-center pb-3">Whose splash art is this?</h3>
-
-      <div
-        className="container d-flex justify-content-center shadow"
-        id="spriteContainer"
-      >
-        <img
-          src={`data:image/webp;base64,${sprite}`}
-          style={{
-            filter: `blur(1.0em) ${isMonochrome ? "grayscale(1)" : ""}`,
-            transform: `${randomRotate ? "rotate(180deg)" : ""}`,
-          }}
-          className="rounded p-4"
-          id="spriteImg"
-          alt="Champion splash art."
-          draggable="false"
+      <Helmet>
+        <title>
+          Infiniteloldle - Ultimate LoL quiz - Champion guessing game
+        </title>
+        <meta
+          name="description"
+          content="The ultimate quiz game for League of Legends enthusiasts. Guess League of Legends champions infinitely."
         />
-      </div>
+      </Helmet>
 
-      <div className="d-flex justify-content-center mt-4 pt-3 mb-3">
+      <h3 className="text-center pb-3">Start guessing your champion</h3>
+
+      <div className="d-flex justify-content-center mt-4 mb-3">
         <form
-          className="form-control row g-3 mb-2"
+          className="form-control row g-3 mb-4"
           onSubmit={Guess}
           id="guess-form"
         >
@@ -198,12 +162,14 @@ export default function Game() {
 
           <div className="d-flex justify-content-evenly">
             {correctGuess ? (
-              <button
-                className="btn btn-light mb-3 mt-1 min-vw-25"
-                onClick={Restart}
-              >
-                Next
-              </button>
+              <>
+                <button
+                  className="btn btn-outline-dark mb-3 mt-1 min-vw-25"
+                  onClick={Restart}
+                >
+                  Next
+                </button>
+              </>
             ) : (
               <button className="btn btn-dark mb-3 mt-1 min-vw-25">
                 Guess
@@ -212,7 +178,7 @@ export default function Game() {
             {!correctGuess && guesses.length >= 10 ? (
               <button
                 className="btn btn-dark mb-3 mt-1 min-vw-25"
-                onClick={() => Reroll("splash")}
+                onClick={() => Reroll("champion")}
               >
                 Reroll
               </button>
@@ -223,13 +189,24 @@ export default function Game() {
         </form>
       </div>
 
-      <div id="championsImgs" className="container">
+      {champions.length > 0 ? <Titles /> : ""}
+
+      <div id="champions">
         {champions.map((champ) => (
-          <ChampionImg
-            key={champ[0].key}
-            championKey={champ[0]}
-            isCorrect={champ[1]}
+          <ChampionDetails
+            key={champ[0].championKey}
+            championKey={champ[0].championKey}
+            gender={champ[0].gender}
+            genre={champ[0].genre}
+            resource={champ[0].resource}
+            rangeTypes={champ[0].rangeType}
+            positions={champ[0].position}
+            releaseYear={champ[0].releaseYear}
+            regions={champ[0].region}
+            damageType={champ[0].damageType}
+            similarites={champ[1]}
             isColorBlindMode={isColorBlindMode}
+            hideResource={hideResource}
           />
         ))}
       </div>
@@ -237,8 +214,8 @@ export default function Game() {
       {correctGuess ? (
         <Victory
           id="victory"
-          championKey={champions[0][0]}
-          champion={currentGuess}
+          championKey={champions[0][0].championKey}
+          champion={champions[0][0].guessedChampion}
           tries={guesses.length}
           title={title}
         />
