+++ frontend/src/views/oldItem.js
@@ -1,5 +1,5 @@
 import React from "react";
-import OldItemGame from "../components/oldItemGame";
+import OldItemGame from "../components/games/oldItemGame";
 import { Helmet } from "react-helmet";
 
 export default function Splash() {
